remove ProgramComponent and make the number of rows in the table dynamic
