+++ stwo_cairo_prover/crates/cairo-air/src/components/constraints_regression_test_values.rs
@@ -198,11 +198,8 @@ pub const VERIFY_BITWISE_XOR_8: Expect = expect![[r#"
 pub const VERIFY_BITWISE_XOR_9: Expect = expect![[r#"
     (841088250 + 1499053200i) + (1122584495 + 937798316i)u
 "#]];
-pub const PROGRAM_COMPONENT: Expect = expect![[r#"
-    (896679863 + 19294025i) + (835905729 + 389684619i)u
-"#]];
 pub const VERIFY_PROGRAM: Expect = expect![[r#"
-    (315471585 + 1243757974i) + (863245459 + 54340856i)u
+    (1104920588 + 1143589918i) + (402921309 + 10191450i)u
 "#]];
 pub const VERIFY_BITWISE_XOR_12: Expect = expect![[r#"
     (1665543176 + 800367929i) + (1406064337 + 277486427i)u
