+++ stwo_cairo_prover/crates/cairo-air/src/components/mod.rs
@@ -86,5 +86,4 @@ pub(crate) fn indented_component_display<E: FrameworkEval>(
         .map(|line| format!("\t{line}"))
         .join("\n")
 }
-pub mod program_component;
 pub mod verify_program;
