+++ stwo_cairo_prover/crates/cairo-air/src/components/prelude.rs
@@ -4,6 +4,7 @@ pub use stwo::core::fields::m31::M31;
 pub use stwo::core::fields::qm31::{SecureField, SECURE_EXTENSION_DEGREE};
 pub use stwo::core::pcs::TreeVec;
 pub use stwo_cairo_common::preprocessed_columns::preprocessed_trace::{PreProcessedColumn, Seq};
+pub use stwo_cairo_common::preprocessed_columns::program::ProgramColumn;
 pub use stwo_cairo_serialize::{CairoDeserialize, CairoSerialize};
 pub use stwo_constraint_framework::preprocessed_columns::PreProcessedColumnId;
 pub use stwo_constraint_framework::{EvalAtRow, FrameworkComponent, FrameworkEval, RelationEntry};
