+++ stwo_cairo_prover/crates/prover/src/utils.rs
@@ -117,9 +117,6 @@ pub fn cairo_provers(components: &CairoComponents) -> Vec<&dyn ComponentProver<S
     if let Some(component) = &components.verify_program {
         vec.push(component as &dyn ComponentProver<SimdBackend>);
     }
-    if let Some(component) = &components.program_component {
-        vec.push(component as &dyn ComponentProver<SimdBackend>);
-    }
     if let Some(component) = &components.pedersen_aggregator_window_bits_18 {
         vec.push(component as &dyn ComponentProver<SimdBackend>);
     }
