+++ stwo_cairo_prover/crates/prover/src/witness/fast_deduction/mod.rs
@@ -2,4 +2,3 @@ pub mod blake;
 pub mod ec_op;
 pub mod pedersen;
 pub mod poseidon;
-pub mod program;
