+++ stwo_cairo_verifier/crates/cairo_air/src/claims.cairo
@@ -21,13 +21,13 @@ use crate::components::{
     pedersen_aggregator_window_bits_18, pedersen_aggregator_window_bits_9, pedersen_builtin,
     pedersen_builtin_narrow_windows, pedersen_points_table_window_bits_18,
     pedersen_points_table_window_bits_9, poseidon_3_partial_rounds_chain, poseidon_aggregator,
-    poseidon_builtin, poseidon_full_round_chain, poseidon_round_keys, program_component,
-    qm_31_add_mul_opcode, range_check96_builtin, range_check_11, range_check_12, range_check_18,
-    range_check_20, range_check_252_width_27, range_check_3_3_3_3_3, range_check_3_6_6_3,
-    range_check_4_3, range_check_4_4, range_check_4_4_4_4, range_check_6, range_check_7_2_5,
-    range_check_8, range_check_9_9, range_check_builtin, ret_opcode, triple_xor_32,
-    verify_bitwise_xor_12, verify_bitwise_xor_4, verify_bitwise_xor_7, verify_bitwise_xor_8,
-    verify_bitwise_xor_9, verify_instruction, verify_program,
+    poseidon_builtin, poseidon_full_round_chain, poseidon_round_keys, qm_31_add_mul_opcode,
+    range_check96_builtin, range_check_11, range_check_12, range_check_18, range_check_20,
+    range_check_252_width_27, range_check_3_3_3_3_3, range_check_3_6_6_3, range_check_4_3,
+    range_check_4_4, range_check_4_4_4_4, range_check_6, range_check_7_2_5, range_check_8,
+    range_check_9_9, range_check_builtin, ret_opcode, triple_xor_32, verify_bitwise_xor_12,
+    verify_bitwise_xor_4, verify_bitwise_xor_7, verify_bitwise_xor_8, verify_bitwise_xor_9,
+    verify_instruction, verify_program,
 };
 use crate::utils::tree_array_concat_cols;
 use crate::{ChannelTrait, PublicDataTrait, components};
@@ -73,7 +73,6 @@ pub struct CairoClaim {
     pub ec_op_builtin: Option<components::ec_op_builtin::Claim>,
     pub partial_ec_mul_generic: Option<components::partial_ec_mul_generic::Claim>,
     pub verify_program: Option<components::verify_program::Claim>,
-    pub program_component: Option<components::program_component::Claim>,
     pub pedersen_aggregator_window_bits_18: Option<
         components::pedersen_aggregator_window_bits_18::Claim,
     >,
@@ -228,9 +227,6 @@ pub impl CairoClaimImpl of ClaimTrait<CairoClaim> {
             log_sizes_list.append(claim.log_sizes());
         }
         if let Some(claim) = self.verify_program {
-            log_sizes_list.append(claim.log_sizes());
-        }
-        if let Some(claim) = self.program_component {
             log_sizes_list.append(claim.log_sizes());
         }
         if let Some(claim) = self.pedersen_aggregator_window_bits_18 {
@@ -749,13 +745,6 @@ pub impl CairoClaimFlattenImpl of CairoClaimFlattenTrait {
             component_log_sizes.append(0_u32);
             component_enable_bits.append(false);
         }
-        if let Some(_c) = self.program_component {
-            component_log_sizes.append(program_component::LOG_SIZE);
-            component_enable_bits.append(true);
-        } else {
-            component_log_sizes.append(0_u32);
-            component_enable_bits.append(false);
-        }
         if let Some(c) = self.pedersen_aggregator_window_bits_18 {
             component_log_sizes.append(*c.log_size);
             component_enable_bits.append(true);
@@ -1035,7 +1024,6 @@ pub struct CairoInteractionClaim {
     pub ec_op_builtin: Option<components::ec_op_builtin::InteractionClaim>,
     pub partial_ec_mul_generic: Option<components::partial_ec_mul_generic::InteractionClaim>,
     pub verify_program: Option<components::verify_program::InteractionClaim>,
-    pub program_component: Option<components::program_component::InteractionClaim>,
     pub pedersen_aggregator_window_bits_18: Option<
         components::pedersen_aggregator_window_bits_18::InteractionClaim,
     >,
@@ -1279,11 +1267,6 @@ pub impl CairoInteractionClaimImpl of CairoInteractionClaimTrace {
             claimed_sums.append(Zero::zero());
         }
         if let Some(c) = self.verify_program {
-            claimed_sums.append(*c.claimed_sum);
-        } else {
-            claimed_sums.append(Zero::zero());
-        }
-        if let Some(c) = self.program_component {
             claimed_sums.append(*c.claimed_sum);
         } else {
             claimed_sums.append(Zero::zero());
@@ -1571,9 +1554,6 @@ pub fn lookup_sum(
         sum += *interaction_claim.claimed_sum;
     }
     if let Some(interaction_claim) = interaction_claim.verify_program {
-        sum += *interaction_claim.claimed_sum;
-    }
-    if let Some(interaction_claim) = interaction_claim.program_component {
         sum += *interaction_claim.claimed_sum;
     }
     if let Some(interaction_claim) = interaction_claim.pedersen_aggregator_window_bits_18 {
