--- stwo_cairo_verifier/crates/cairo_air/src/components/program_component.cairo
@@ -1,176 +1,0 @@
-// This file was created by the AIR team.
-
-use crate::prelude::*;
-
-pub const N_TRACE_COLUMNS: usize = 1;
-pub const LOG_SIZE: u32 = 12;
-
-#[derive(Drop, Serde, Copy)]
-pub struct Claim {}
-
-pub impl ClaimImpl of ClaimTrait<Claim> {
-    fn log_sizes(self: @Claim) -> TreeArray<Span<u32>> {
-        let log_size = LOG_SIZE;
-        let preprocessed_log_sizes = array![log_size].span();
-        let trace_log_sizes = [log_size; N_TRACE_COLUMNS].span();
-        let interaction_log_sizes = [log_size; 4].span();
-        array![preprocessed_log_sizes, trace_log_sizes, interaction_log_sizes]
-    }
-
-    fn mix_into(self: @Claim, ref channel: Channel) {}
-
-    fn accumulate_relation_uses(self: @Claim, ref relation_uses: RelationUsesDict) {}
-}
-
-#[derive(Drop, Serde, Copy)]
-pub struct InteractionClaim {
-    pub claimed_sum: QM31,
-}
-
-#[generate_trait]
-pub impl InteractionClaimImpl of InteractionClaimTrait {
-    fn mix_into(self: @InteractionClaim, ref channel: Channel) {
-        channel.mix_felts([*self.claimed_sum].span());
-    }
-}
-
-
-#[derive(Drop)]
-pub struct Component {
-    pub claim: Claim,
-    pub interaction_claim: InteractionClaim,
-    pub common_lookup_elements: CommonLookupElements,
-}
-
-pub impl NewComponentImpl of NewComponent<Component> {
-    type Claim = Claim;
-    type InteractionClaim = InteractionClaim;
-
-    fn new(
-        claim: @Claim,
-        interaction_claim: @InteractionClaim,
-        common_lookup_elements: @CommonLookupElements,
-    ) -> Component {
-        Component {
-            claim: *claim,
-            interaction_claim: *interaction_claim,
-            common_lookup_elements: common_lookup_elements.clone(),
-        }
-    }
-}
-
-pub impl CairoComponentImpl of CairoComponent<Component> {
-    fn evaluate_constraints_at_point(
-        self: @Component,
-        ref sum: QM31,
-        ref preprocessed_mask_values: PreprocessedMaskValues,
-        ref trace_mask_values: ColumnSpan<Span<QM31>>,
-        ref interaction_trace_mask_values: ColumnSpan<Span<QM31>>,
-        random_coeff: QM31,
-    ) {
-        let log_size = LOG_SIZE;
-        let claimed_sum = *self.interaction_claim.claimed_sum;
-        let column_size = m31(pow2(log_size));
-        let mut program_component_sum_0: QM31 = Zero::zero();
-        let seq_12 = preprocessed_mask_values.get_and_mark_used(SEQ_12_IDX);
-        let curr_program_0 = preprocessed_mask_values.get_and_mark_used(CURR_PROGRAM_0_IDX);
-        let curr_program_1 = preprocessed_mask_values.get_and_mark_used(CURR_PROGRAM_1_IDX);
-        let curr_program_2 = preprocessed_mask_values.get_and_mark_used(CURR_PROGRAM_2_IDX);
-        let curr_program_3 = preprocessed_mask_values.get_and_mark_used(CURR_PROGRAM_3_IDX);
-        let curr_program_4 = preprocessed_mask_values.get_and_mark_used(CURR_PROGRAM_4_IDX);
-        let curr_program_5 = preprocessed_mask_values.get_and_mark_used(CURR_PROGRAM_5_IDX);
-        let curr_program_6 = preprocessed_mask_values.get_and_mark_used(CURR_PROGRAM_6_IDX);
-        let curr_program_7 = preprocessed_mask_values.get_and_mark_used(CURR_PROGRAM_7_IDX);
-        let curr_program_8 = preprocessed_mask_values.get_and_mark_used(CURR_PROGRAM_8_IDX);
-        let curr_program_9 = preprocessed_mask_values.get_and_mark_used(CURR_PROGRAM_9_IDX);
-        let curr_program_10 = preprocessed_mask_values.get_and_mark_used(CURR_PROGRAM_10_IDX);
-        let curr_program_11 = preprocessed_mask_values.get_and_mark_used(CURR_PROGRAM_11_IDX);
-        let curr_program_12 = preprocessed_mask_values.get_and_mark_used(CURR_PROGRAM_12_IDX);
-        let curr_program_13 = preprocessed_mask_values.get_and_mark_used(CURR_PROGRAM_13_IDX);
-        let curr_program_14 = preprocessed_mask_values.get_and_mark_used(CURR_PROGRAM_14_IDX);
-        let curr_program_15 = preprocessed_mask_values.get_and_mark_used(CURR_PROGRAM_15_IDX);
-        let curr_program_16 = preprocessed_mask_values.get_and_mark_used(CURR_PROGRAM_16_IDX);
-        let curr_program_17 = preprocessed_mask_values.get_and_mark_used(CURR_PROGRAM_17_IDX);
-        let curr_program_18 = preprocessed_mask_values.get_and_mark_used(CURR_PROGRAM_18_IDX);
-        let curr_program_19 = preprocessed_mask_values.get_and_mark_used(CURR_PROGRAM_19_IDX);
-        let curr_program_20 = preprocessed_mask_values.get_and_mark_used(CURR_PROGRAM_20_IDX);
-        let curr_program_21 = preprocessed_mask_values.get_and_mark_used(CURR_PROGRAM_21_IDX);
-        let curr_program_22 = preprocessed_mask_values.get_and_mark_used(CURR_PROGRAM_22_IDX);
-        let curr_program_23 = preprocessed_mask_values.get_and_mark_used(CURR_PROGRAM_23_IDX);
-        let curr_program_24 = preprocessed_mask_values.get_and_mark_used(CURR_PROGRAM_24_IDX);
-        let curr_program_25 = preprocessed_mask_values.get_and_mark_used(CURR_PROGRAM_25_IDX);
-        let curr_program_26 = preprocessed_mask_values.get_and_mark_used(CURR_PROGRAM_26_IDX);
-        let curr_program_27 = preprocessed_mask_values.get_and_mark_used(CURR_PROGRAM_27_IDX);
-
-        let [program_segment_multiplicity]: [Span<QM31>; 1] = (*trace_mask_values
-            .multi_pop_front()
-            .unwrap())
-            .unbox();
-        let [program_segment_multiplicity]: [QM31; 1] = (*program_segment_multiplicity
-            .try_into()
-            .unwrap())
-            .unbox();
-        core::internal::revoke_ap_tracking();
-
-        program_component_sum_0 = self
-            .common_lookup_elements
-            .combine_qm31(
-                [
-                    qm31_const::<1942035206, 0, 0, 0>(), seq_12, curr_program_0, curr_program_1,
-                    curr_program_2, curr_program_3, curr_program_4, curr_program_5, curr_program_6,
-                    curr_program_7, curr_program_8, curr_program_9, curr_program_10,
-                    curr_program_11, curr_program_12, curr_program_13, curr_program_14,
-                    curr_program_15, curr_program_16, curr_program_17, curr_program_18,
-                    curr_program_19, curr_program_20, curr_program_21, curr_program_22,
-                    curr_program_23, curr_program_24, curr_program_25, curr_program_26,
-                    curr_program_27,
-                ]
-                    .span(),
-            );
-
-        lookup_constraints(
-            ref sum,
-            random_coeff,
-            claimed_sum,
-            program_segment_multiplicity,
-            column_size,
-            ref interaction_trace_mask_values,
-            program_component_sum_0,
-        );
-    }
-}
-
-
-fn lookup_constraints(
-    ref sum: QM31,
-    random_coeff: QM31,
-    claimed_sum: QM31,
-    program_segment_multiplicity: QM31,
-    column_size: M31,
-    ref interaction_trace_mask_values: ColumnSpan<Span<QM31>>,
-    program_component_sum_0: QM31,
-) {
-    let [trace_2_col0, trace_2_col1, trace_2_col2, trace_2_col3]: [Span<QM31>; 4] =
-        (*interaction_trace_mask_values
-        .multi_pop_front()
-        .unwrap())
-        .unbox();
-
-    let [trace_2_col0_neg1, trace_2_col0]: [QM31; 2] = (*trace_2_col0.try_into().unwrap()).unbox();
-    let [trace_2_col1_neg1, trace_2_col1]: [QM31; 2] = (*trace_2_col1.try_into().unwrap()).unbox();
-    let [trace_2_col2_neg1, trace_2_col2]: [QM31; 2] = (*trace_2_col2.try_into().unwrap()).unbox();
-    let [trace_2_col3_neg1, trace_2_col3]: [QM31; 2] = (*trace_2_col3.try_into().unwrap()).unbox();
-
-    core::internal::revoke_ap_tracking();
-
-    let constraint_quotient = (((QM31Impl::from_partial_evals(
-        [trace_2_col0, trace_2_col1, trace_2_col2, trace_2_col3],
-    )
-        - QM31Impl::from_partial_evals(
-            [trace_2_col0_neg1, trace_2_col1_neg1, trace_2_col2_neg1, trace_2_col3_neg1],
-        )
-        + (claimed_sum * (column_size.inverse().into())))
-        * program_component_sum_0)
-        + program_segment_multiplicity);
-    sum = sum * random_coeff + constraint_quotient;
-}
