+++ stwo_cairo_verifier/crates/cairo_air/src/components/sample_evaluations.cairo
@@ -1,4 +1,4 @@
-// AIR version 64d4253e
+// AIR version 1160acc8
 use stwo_verifier_core::fields::m31::M31;
 pub const ADD_AP_OPCODE_SAMPLE_EVAL_RESULT: [M31; 4] = [
     M31 { inner: 1435814162 }, M31 { inner: 1618992457 }, M31 { inner: 840348268 },
@@ -172,10 +172,6 @@ pub const POSEIDON_ROUND_KEYS_SAMPLE_EVAL_RESULT: [M31; 4] = [
     M31 { inner: 1144238612 }, M31 { inner: 1473001273 }, M31 { inner: 1284604800 },
     M31 { inner: 1049928592 },
 ];
-pub const PROGRAM_COMPONENT_SAMPLE_EVAL_RESULT: [M31; 4] = [
-    M31 { inner: 1831889950 }, M31 { inner: 1158688806 }, M31 { inner: 167508377 },
-    M31 { inner: 1244309204 },
-];
 pub const QM_31_ADD_MUL_OPCODE_SAMPLE_EVAL_RESULT: [M31; 4] = [
     M31 { inner: 1161283513 }, M31 { inner: 1571148078 }, M31 { inner: 95782107 },
     M31 { inner: 852512197 },
@@ -277,6 +273,6 @@ pub const VERIFY_INSTRUCTION_SAMPLE_EVAL_RESULT: [M31; 4] = [
     M31 { inner: 1058351204 },
 ];
 pub const VERIFY_PROGRAM_SAMPLE_EVAL_RESULT: [M31; 4] = [
-    M31 { inner: 1500873141 }, M31 { inner: 1343289352 }, M31 { inner: 344861752 },
-    M31 { inner: 388001067 },
+    M31 { inner: 1253286041 }, M31 { inner: 1712881693 }, M31 { inner: 947747979 },
+    M31 { inner: 1770681791 },
 ];
